refactor(content): WKC layout — Phase 12 (test alignment)
Phase 12 of the WKC migration. Realigns three spec file locations to
mirror the renames and moves from earlier phases.

Renames / moves:

- test/unit/logic/cluster-component.spec.ts ->
  test/unit/logic/peer-cluster.spec.ts
  (mirrors Phase 7's logic/cluster -> logic/peer-cluster rename)
- test/unit/service/validations/server.spec.ts ->
  test/unit/logic/server-validator.spec.ts
  (mirrors Phase 4's service/validations/server -> logic/server-validator)
- test/unit/service/synchronization/batchDeployer.spec.ts ->
  test/unit/logic/batch-deployer.spec.ts
  (mirrors Phase 7's service/synchronization/batchDeployer -> logic/batch-deployer)

Relative-import depth updated where the move changed levels (the spec
files in test/unit/service/X/Y/ used '../../../../src/...' four levels
up; their new homes in test/unit/logic/ are three levels and use
'../../../src/...').

After the moves, test/unit/service/ contains only pointers/ — the
NoOpPointerManager test helper and pointers.spec.ts still live there
because their paths are referenced by the deployer and active-entities
specs via relative paths; moving them is a follow-up that requires
more invasive sed.

What this PR deliberately does NOT do:

- No mock factory rewrites. test/mocks/ continues to mock components
  by their exported types; existing patterns work against the new
  component shapes without changes.
- No new component-level unit tests. The plan flagged "add unit tests
  for the deployment-service transaction coordinator and the
  content-validator strategy selection" as Phase 12 work, but those
  are net-new test authoring rather than reorganization. Defer.
- No integration-test reorganization. The integration suite paths
  reference the new src/ layout via the existing imports and run
  end-to-end against a real Postgres; their location under
  test/integration/<area>/ is independently meaningful.

tsc clean, unit suite green (213 tests, no diff).
